+++ pkg/R/rgeos_topology.R
@@ -90,6 +90,8 @@ TopologyFunc = function(spgeom, id, byid, func) {
         x <- .Call("rgeos_makevalid", .RGEOS_HANDLE, spgeom, id, byid, PACKAGE="rgeos")
     else if (func == "rgeos_coverageunion")
         x <- .Call("rgeos_coverageunion", .RGEOS_HANDLE, spgeom, id, byid, PACKAGE="rgeos")
+    else if (func == "rgeos_minimumrotatedrectangle")
+        x <- .Call("rgeos_minimumrotatedrectangle", .RGEOS_HANDLE, spgeom, id, byid, PACKAGE="rgeos")
     else stop("no such function:", func)
 
 
@@ -171,6 +173,64 @@ gCoverageUnion = function(spgeom, byid=FALSE, id = NULL) {
     return( TopologyFunc(spgeom, id, byid, "rgeos_coverageunion") ) 
 
 }
+
+gMinumumRotatedRectangle <- function(spgeom, byid=FALSE, id = NULL) {
+
+    if (version_GEOS0() < "3.9.0")
+        stop("No MinumumRotatedRectangle in this version of GEOS")
+    return( TopologyFunc(spgeom, id, byid, "rgeos_minimumrotatedrectangle") ) 
+
+}
+
+
+gMaximumInscribedCircle <- function(spgeom, byid=FALSE, id = NULL, tol=.Machine$double.eps^(1/2), checkValidity=NULL) {    
+
+    if (version_GEOS0() < "3.9.0")
+        stop("No MaximumInscribedCircle in this version of GEOS")
+
+    if (!inherits(spgeom,"SpatialPolygons"))
+        stop("Invalid geometry, may only be applied to polygons")
+
+    stopifnot(is.logical(byid))
+    byid = as.logical(byid)
+    if (is.na(byid)) 
+        stop("Invalid value for byid, must be logical")
+
+    if (is.null(checkValidity)) checkValidity <- get_RGEOS_CheckValidity()
+    if(checkValidity > 0L) {
+        val1 <- isTRUE(all(gIsValid(spgeom, byid=TRUE)))
+        nm_1 <- deparse(substitute(spgeom))
+        if (!val1) message(nm_1, " is invalid")
+        if (checkValidity == 1L && !all(val1)) warning("Invalid objects found; consider using set_RGEOS_CheckValidity(2L)")
+        if (checkValidity == 2L) {
+            if (!val1) {
+                spgeom <- gBuffer(spgeom, byid=TRUE, width=0.0)
+                message("Attempting to make ", nm_1, " valid by zero-width buffering")
+                if (!isTRUE(all(gIsValid(spgeom, byid=TRUE)))) stop("Zero width buffer repair of spgeom1 failed")
+            }
+        }
+    }
+
+    spgeom <- as(spgeom, "SpatialPolygons")
+    curids = unique(row.names(spgeom))
+    if (is.null(id)) {
+        if (byid)   id = curids
+        else        id = "1"
+    }
+    id = as.character(id)
+
+#    if (any(is.na(id))) stop("No NAs permitted in id")
+
+    if (get_do_poly_check() && notAllComments(spgeom)) spgeom <- createSPComment(spgeom)
+
+    .Call("rgeos_maximuminscribedcircle", .RGEOS_HANDLE,
+        spgeom, id, byid, tol, PACKAGE="rgeos")
+
+
+#    return( TopologyFunc(spgeom, id, byid, tol, "rgeos_maximuminscribedcircle") ) 
+
+}
+
 
 gUnaryUnion = function(spgeom, id = NULL, checkValidity=NULL) {