+++ pkg/NEWS.md
@@ -4,6 +4,8 @@
 
 * update Windows static library to GEOS 3.9
 
+* add gMinumumRotatedRectangle() and gMaximumInscribedCircle() for GEOS 3.9
+
 # Version 0.5-5 (2020-09-07, rev. 634-640)
 
 * add gMakeValid() for GEOS 3.8